+++ common/src/main/kotlin/com/cobblemon/mod/common/client/gui/pc/PCGUI.kt
@@ -35,13 +35,16 @@ import net.minecraft.sound.SoundEvent
 import net.minecraft.text.Text
 
 class PCGUI(
-    private val pc: ClientPC,
-    private val party: ClientParty
+    val pc: ClientPC,
+    val party: ClientParty,
+    val configuration: PCGUIConfiguration
 ) : Screen(Text.translatable("cobblemon.ui.pc.title")) {
 
     companion object {
         const val BASE_WIDTH = 349
         const val BASE_HEIGHT = 205
+        const val RIGHT_PANEL_WIDTH = 82
+        const val RIGHT_PANEL_HEIGHT = 169
         const val TYPE_SPACER_WIDTH = 128
         const val TYPE_SPACER_HEIGHT = 12
         const val PC_SPACER_WIDTH = 342
@@ -72,13 +75,7 @@ class PCGUI(
         val y = (height - BASE_HEIGHT) / 2
 
         // Add Exit Button
-        this.addDrawableChild(
-            ExitButton(pX = x + 320, pY = y + 186) {
-                playSound(CobblemonSounds.PC_OFF)
-                MinecraftClient.getInstance().setScreen(null)
-                UnlinkPlayerFromPCPacket().sendToServer()
-            }
-        )
+        this.addDrawableChild(ExitButton(pX = x + 320, pY = y + 186) { configuration.exitFunction(this) })
 
         // Add Forward Button
         this.addDrawableChild(
@@ -408,6 +405,24 @@ class PCGUI(
         }
     }
 
+    fun closeNormally(unlink: Boolean = true) {
+        playSound(CobblemonSounds.PC_OFF)
+        MinecraftClient.getInstance().setScreen(null)
+        if (unlink) {
+            UnlinkPlayerFromPCPacket().sendToServer()
+        }
+    }
+
+    override fun mouseScrolled(mouseX: Double, mouseY: Double, amount: Double): Boolean {
+        if (storageWidget.pastureWidget != null) storageWidget.pastureWidget!!.pastureScrollList.mouseScrolled(mouseX, mouseY, amount)
+        return children().any { it.mouseScrolled(mouseX, mouseY, amount) }
+    }
+
+    override fun mouseDragged(mouseX: Double, mouseY: Double, button: Int, deltaX: Double, deltaY: Double): Boolean {
+        if (storageWidget.pastureWidget != null) storageWidget.pastureWidget!!.pastureScrollList.mouseDragged(mouseX, mouseY, button, deltaX, deltaY)
+        return super.mouseDragged(mouseX, mouseY, button, deltaX, deltaY)
+    }
+
     override fun keyPressed(keyCode: Int, scanCode: Int, modifiers: Int): Boolean {
         when (keyCode) {
             InputUtil.GLFW_KEY_ESCAPE -> {
@@ -442,7 +457,7 @@ class PCGUI(
         if (ticksElapsed % delayFactor == 0) selectPointerOffsetY += if (selectPointerOffsetIncrement) 1 else -1
     }
 
-    private fun playSound(soundEvent: SoundEvent) {
+    fun playSound(soundEvent: SoundEvent) {
         MinecraftClient.getInstance().soundManager.play(PositionedSoundInstance.master(soundEvent, 1.0F))
     }