+++ common/src/main/kotlin/com/cobblemon/mod/common/client/gui/pc/StorageWidget.kt
@@ -16,6 +16,8 @@ import com.cobblemon.mod.common.api.storage.party.PartyPosition
 import com.cobblemon.mod.common.api.storage.pc.PCPosition
 import com.cobblemon.mod.common.api.text.bold
 import com.cobblemon.mod.common.client.CobblemonResources
+import com.cobblemon.mod.common.client.gui.pasture.PasturePCGUIConfiguration
+import com.cobblemon.mod.common.client.gui.pasture.PastureWidget
 import com.cobblemon.mod.common.client.gui.summary.widgets.SoundlessWidget
 import com.cobblemon.mod.common.client.render.drawScaledText
 import com.cobblemon.mod.common.client.settings.ServerSettings
@@ -58,6 +60,7 @@ class StorageWidget(
         const val BOX_SLOT_PADDING = 2
         const val PARTY_SLOT_PADDING = 6
 
+        private val partyPanelResource = cobblemonResource("textures/gui/pc/party_panel.png")
         private val screenOverlayResource = cobblemonResource("textures/gui/pc/pc_screen_overlay.png")
     }
 
@@ -66,6 +69,8 @@ class StorageWidget(
     private val releaseButton: ReleaseButton
     private val releaseYesButton: ReleaseConfirmButton
     private val releaseNoButton: ReleaseConfirmButton
+
+    var pastureWidget: PastureWidget? = null
 
     var displayConfirmRelease = false
     var screenLoaded = false
@@ -134,6 +139,10 @@ class StorageWidget(
                 }
             }
         )
+
+        if (pcGui.configuration is PasturePCGUIConfiguration) {
+            this.pastureWidget = PastureWidget(pcGui.configuration, x + 182, y - 19)
+        }
     }
 
     fun canDeleteSelected(): Boolean {
@@ -168,30 +177,32 @@ class StorageWidget(
         }
 
         // Party Slots
-        for (partyIndex in 0..5) {
-            var partyX = x + PARTY_SLOT_START_OFFSET_X
-            var partyY = y + PARTY_SLOT_START_OFFSET_Y
+        if (pcGui.configuration.showParty) {
+            for (partyIndex in 0..5) {
+                var partyX = x + PARTY_SLOT_START_OFFSET_X
+                var partyY = y + PARTY_SLOT_START_OFFSET_Y
 
-            if (partyIndex > 0) {
-                val isEven = partyIndex % 2 == 0
-                val offsetIndex = (partyIndex - (if (isEven) 0 else 1)) / 2
-                val offsetX = if (isEven) 0 else (StorageSlot.SIZE + PARTY_SLOT_PADDING)
-                val offsetY = if (isEven) 0 else 8
+                if (partyIndex > 0) {
+                    val isEven = partyIndex % 2 == 0
+                    val offsetIndex = (partyIndex - (if (isEven) 0 else 1)) / 2
+                    val offsetX = if (isEven) 0 else (StorageSlot.SIZE + PARTY_SLOT_PADDING)
+                    val offsetY = if (isEven) 0 else 8
 
-                partyX += offsetX
-                partyY += ((StorageSlot.SIZE + PARTY_SLOT_PADDING) * offsetIndex) + offsetY
-            }
+                    partyX += offsetX
+                    partyY += ((StorageSlot.SIZE + PARTY_SLOT_PADDING) * offsetIndex) + offsetY
+                }
 
-            PartyStorageSlot(
-                x = partyX,
-                y = partyY,
-                parent = this,
-                party = party,
-                position = PartyPosition(partyIndex),
-                onPress = { this.onStorageSlotClicked(it) }
-            ).also { widget ->
-                this.addWidget(widget)
-                this.partySlots.add(widget)
+                PartyStorageSlot(
+                    x = partyX,
+                    y = partyY,
+                    parent = this,
+                    party = party,
+                    position = PartyPosition(partyIndex),
+                    onPress = { this.onStorageSlotClicked(it) }
+                ).also { widget ->
+                    this.addWidget(widget)
+                    this.partySlots.add(widget)
+                }
             }
         }
     }
@@ -207,31 +218,43 @@ class StorageWidget(
 
     override fun renderButton(matrices: MatrixStack, mouseX: Int, mouseY: Int, delta: Float) {
         // Party  Label
-        drawScaledText(
-            matrixStack = matrices,
-            font = CobblemonResources.DEFAULT_LARGE,
-            text = lang("ui.party").bold(),
-            x = x + 213,
-            y = y - 15.5,
-            centered = true,
-            shadow = true
-        )
+        if (pcGui.configuration.showParty) {
+            blitk(
+                matrixStack = matrices,
+                texture = partyPanelResource,
+                x = x + 182,
+                y = y - 19,
+                width = PCGUI.RIGHT_PANEL_WIDTH,
+                height = PCGUI.RIGHT_PANEL_HEIGHT
+            )
 
-        if (canDeleteSelected() && displayConfirmRelease) {
             drawScaledText(
                 matrixStack = matrices,
                 font = CobblemonResources.DEFAULT_LARGE,
-                text = lang("ui.pc.release").bold(),
-                x = x + 223,
-                y = y + 119,
-                centered = true
+                text = lang("ui.party").bold(),
+                x = x + 213,
+                y = y - 15.5,
+                centered = true,
+                shadow = true
             )
+
+
+            if (canDeleteSelected() && displayConfirmRelease) {
+                drawScaledText(
+                    matrixStack = matrices,
+                    font = CobblemonResources.DEFAULT_LARGE,
+                    text = lang("ui.pc.release").bold(),
+                    x = x + 223,
+                    y = y + 119,
+                    centered = true
+                )
+            }
+
+            this.releaseButton.render(matrices, mouseX, mouseY, delta)
+            this.releaseYesButton.render(matrices, mouseX, mouseY, delta)
+            this.releaseNoButton.render(matrices, mouseX, mouseY, delta)
         }
 
-        this.releaseButton.render(matrices, mouseX, mouseY, delta)
-        this.releaseYesButton.render(matrices, mouseX, mouseY, delta)
-        this.releaseNoButton.render(matrices, mouseX, mouseY, delta)
-        
         // Screen Overlay
         blitk(
             matrixStack = matrices,
@@ -254,12 +277,18 @@ class StorageWidget(
             if (pcGui.ticksElapsed >= 10)  screenLoaded = true
         }
 
-        this.partySlots.forEach { slot ->
-            slot.render(matrices, mouseX, mouseY, delta)
-            val pokemon = slot.getPokemon()
-            if (grabbedSlot == null && slot.isHovered(mouseX, mouseY)
-                && pokemon != null && pokemon != pcGui.previewPokemon) pcGui.setPreviewPokemon(pokemon)
+        // Party slots
+        if (pcGui.configuration.showParty) {
+            this.partySlots.forEach { slot ->
+                slot.render(matrices, mouseX, mouseY, delta)
+                val pokemon = slot.getPokemon()
+                if (grabbedSlot == null && slot.isHovered(mouseX, mouseY)
+                    && pokemon != null && pokemon != pcGui.previewPokemon
+                ) pcGui.setPreviewPokemon(pokemon)
+            }
         }
+
+        pastureWidget?.renderButton(matrices, mouseX, mouseY, delta)
 
         grabbedSlot?.render(matrices, mouseX, mouseY, delta)
     }
@@ -272,6 +301,7 @@ class StorageWidget(
             if (releaseButton.isHovered(pMouseX, pMouseY)) releaseButton.mouseClicked(pMouseX, pMouseY, pButton)
         }
 
+        pastureWidget?.mouseClicked(pMouseX, pMouseY, pButton)
         return super.mouseClicked(pMouseX, pMouseY, pButton)
     }
 
@@ -288,7 +318,6 @@ class StorageWidget(
     }
 
     private fun onStorageSlotClicked(button: ButtonWidget) {
-
         // Check if storage slot
         val clickedPosition = when(button) {
             is BoxStorageSlot -> button.position
@@ -311,6 +340,12 @@ class StorageWidget(
             is BoxStorageSlot -> pc.get(clickedPosition as PCPosition)
             is PartyStorageSlot -> party.get(clickedPosition as PartyPosition)
             else -> null
+        }
+
+        val selectOverride = pcGui.configuration.selectOverride
+        if (selectOverride != null) {
+            selectOverride(pcGui, clickedPosition, clickedPokemon)
+            return
         }
 
         if (grabbedSlot == null) {