+++ common/src/main/kotlin/com/cobblemon/mod/common/net/messages/client/pokemon/update/ShinyUpdatePacket.kt
@@ -12,7 +12,7 @@ import com.cobblemon.mod.common.pokemon.Pokemon
 import com.cobblemon.mod.common.util.cobblemonResource
 import net.minecraft.network.PacketByteBuf
 
-class ShinyUpdatePacket(pokemon: Pokemon, value: Boolean) : BooleanUpdatePacket<ShinyUpdatePacket>(pokemon, value) {
+class ShinyUpdatePacket(pokemon: () -> Pokemon, value: Boolean) : BooleanUpdatePacket<ShinyUpdatePacket>(pokemon, value) {
     override val id = ID
     override fun set(pokemon: Pokemon, value: Boolean) { pokemon.shiny = value }
     companion object {